backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.8.2
pydantic-settings==2.4.0
motor >= 3.5.1
pandas==2.2.2
scikit-learn==1.4.2
requests==2.32.3
httpx==0.27.0
celery==5.4.0
redis==5.0.7
email-validator==2.2.0
pymongo>=4.5,<5